refactor: #49 パスとディレクトリを調整
