+++ src/app/routes/app/wallet/crypto-wallet/exchange/digital-gift.tsx
@@ -1,11 +1,11 @@
-import { Complete } from "@/features/app/crypto-wallet/components/exchange/complete";
 import { WalletLayout } from "@/features/app/layouts/wallet-layout";
+import { DigitalGift } from "@/features/app/wallet/crypto-wallet/components/exchange/digital-gift";
 
-export const CompleteRoute = () => {
+export const DigitalGiftRoute = () => {
   return (
     <div className="flex flex-col gap-y-5 items-center pt-[30px] px-5">
       <WalletLayout label="Exchange">
-        <Complete />
+        <DigitalGift />
       </WalletLayout>
     </div>
   );
