+++ src/app/routes/app/wallet/crypto-wallet/exchange/token.tsx
@@ -1,11 +1,11 @@
-import { DigitalGift } from "@/features/app/crypto-wallet/components/exchange/digital-gift";
 import { WalletLayout } from "@/features/app/layouts/wallet-layout";
+import { Token } from "@/features/app/wallet/crypto-wallet/components/exchange/token";
 
-export const DigitalGiftRoute = () => {
+export const TokenRoute = () => {
   return (
     <div className="flex flex-col gap-y-5 items-center pt-[30px] px-5">
       <WalletLayout label="Exchange">
-        <DigitalGift />
+        <Token />
       </WalletLayout>
     </div>
   );
