+++ src/app/routes/app/wallet/crypto-wallet/transfer.tsx
@@ -1,11 +1,11 @@
-import { Token } from "@/features/app/crypto-wallet/components/exchange/token";
 import { WalletLayout } from "@/features/app/layouts/wallet-layout";
+import { Transfer } from "@/features/app/wallet/crypto-wallet/components/transfer";
 
-export const TokenRoute = () => {
+export const TransferRoute = () => {
   return (
     <div className="flex flex-col gap-y-5 items-center pt-[30px] px-5">
-      <WalletLayout label="Exchange">
-        <Token />
+      <WalletLayout label="Transfer">
+        <Transfer />
       </WalletLayout>
     </div>
   );
