+++ src/features/app/wallet/crypto-wallet/components/exchange/digital-gift.tsx
@@ -5,17 +5,17 @@ import { Pulldown } from "@/components/ui/form/pulldown";
 import { TextField } from "@/components/ui/form/text-field";
 import { TextFieldPulldown } from "@/components/ui/form/text-field-pulldown";
 import { LinkTabs } from "@/components/ui/link-tabs";
+import { ThanksCard } from "@/features/app/components/thanks-card";
 import { useGetUser } from "@/features/profile/api";
 import { path } from "@/utils/path";
 import { Box, Stack } from "@mui/material";
 import type { FormEvent } from "react";
-import { ThanksCard } from "../../../components/thanks-card";
+import { CardHeader } from "../card-header";
 import {
   DIGITAL_GIFT_OPTIONS,
   REGION_OPTIONS,
   VALUE_OPTIONS,
-} from "../../constants";
-import { CardHeader } from "../card-header";
+} from "../constants";
 
 export const DigitalGift = () => {
   const { user: authorizedUser } = useUser();
@@ -43,15 +43,15 @@ export const DigitalGift = () => {
           linkButtons={[
             {
               label: "Transfer",
-              to: path.get().app.cryptoWallet.transfer,
+              to: path.get().app.wallet.cryptoWallet.transfer,
             },
             {
               label: "Swap",
-              to: path.get().app.cryptoWallet.swap,
+              to: path.get().app.wallet.cryptoWallet.swap,
             },
             {
               label: "Exchange",
-              to: path.get().app.cryptoWallet.exchange.digitalGift,
+              to: path.get().app.wallet.cryptoWallet.exchange.digitalGift,
               isCurrentPage: true,
             },
           ]}
@@ -63,14 +63,14 @@ export const DigitalGift = () => {
             tabs={[
               {
                 label: "Digital Gift",
-                to: path.get().app.cryptoWallet.exchange.digitalGift,
+                to: path.get().app.wallet.cryptoWallet.exchange.digitalGift,
               },
               {
                 label: "Token",
-                to: path.get().app.cryptoWallet.exchange.token,
+                to: path.get().app.wallet.cryptoWallet.exchange.token,
               },
             ]}
-            value={path.get().app.cryptoWallet.exchange.digitalGift}
+            value={path.get().app.wallet.cryptoWallet.exchange.digitalGift}
           />
         </Box>
